Change data type from int to long

+++ MosaicResidentInformationApi.Tests/V1/Helper/TestHelper.cs
@@ -9,7 +9,7 @@ namespace MosaicResidentInformationApi.Tests.V1.Helper
 {
     public static class TestHelper
     {
-        public static Person CreateDatabasePersonEntity(string firstname = null, string lastname = null, int? id = null)
+        public static Person CreateDatabasePersonEntity(string firstname = null, string lastname = null, long? id = null)
         {
             var faker = new Fixture();
             var fp = faker.Build<Person>()
@@ -24,7 +24,7 @@ namespace MosaicResidentInformationApi.Tests.V1.Helper
             return fp;
         }
 
-        public static Address CreateDatabaseAddressForPersonId(int personId, string postcode = null, string address = null)
+        public static Address CreateDatabaseAddressForPersonId(long personId, string postcode = null, string address = null)
         {
             var faker = new Fixture();
 
@@ -39,7 +39,7 @@ namespace MosaicResidentInformationApi.Tests.V1.Helper
             return fa;
         }
 
-        public static TelephoneNumber CreateDatabaseTelephoneNumberForPersonId(int personId)
+        public static TelephoneNumber CreateDatabaseTelephoneNumberForPersonId(long personId)
         {
             var faker = new Fixture();
